+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.0.2
+- Updated button texture
+- Added selection notification
+
 ### v1.0.1
 - Moved position to right
 - requires Node IDs now cuz of some code